+++ packages/app/src/components/Main/Create/index.ts
@@ -1,0 +1,1 @@
+export * from "./Create";
