--- packages/app/src/components/Registry/index.ts
@@ -1,1 +1,0 @@
-export * from "./Registry";
