+++ packages/app/src/type/confg.ts
@@ -1,4 +1,3 @@
-// export "" from ""
 import configs from "../../config.json";
 
 export type LinkKey = keyof typeof configs.links;
