--- packages/app/src/type/txList.ts
@@ -1,3 +1,0 @@
-export interface TXList {
-  [chainId: string]: string[];
-}
